+++ src/util/plant/repository.go
@@ -5,5 +5,5 @@ import "github.com/KitaPDev/fogfarms-server/models"
 type Repository interface {
 	GetPlant(plantID string) *models.Plant
 	GetAllPlants() []models.Plant
-	NewPlant(name string, tds float32, ph float32, lux float32)
+	NewPlant(name string, tds float64, ph float64, lux float64)
 }
